add issues and refactor some files
